# backend/src/test/resources/openapi/api-docs.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Mail Backend API",
    "description" : "REST API for the Tauri desktop mail client — accounts, messages, drafts, folders.",
    "license" : {
      "name" : "Proprietary"
    },
    "version" : "1.0.0"
  },
  "security" : [ {
    "ApiKeyAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/v1/messages/account/{accountId}/send" : {
      "post" : {
        "tags" : [ "Messages — Write" ],
        "summary" : "Send message (async)",
        "description" : "Asynchronously sends a message from the given account. Returns 202 Accepted with a sendId — the client tracks the outcome via the send_completed / send_failed notification stream event.",
        "operationId" : "sendEmail",
        "parameters" : [ {
          "name" : "accountId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "exclusiveMinimum" : 0
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MailRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "503" : {
            "description" : "IMAP/SMTP server unavailable (MAIL_CONNECTION_ERROR).",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "202" : {
            "description" : "Send request accepted; outcome is delivered via the notification stream.",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendAcceptedResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid X-API-KEY.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input (validation error).",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/messages/{stableId}" : {
      "get" : {
        "tags" : [ "Messages — Read" ],
        "summary" : "Message detail",
        "description" : "Returns the full message detail including headers, recipients and attachment list (metadata, not content). For the message body see /content.",
        "operationId" : "getMessageDetail",
        "parameters" : [ {
          "name" : "stableId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "maxLength" : 128,
            "minLength" : 0
          }
        } ],
        "responses" : {
          "503" : {
            "description" : "IMAP/SMTP server unavailable (MAIL_CONNECTION_ERROR).",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MailDetailResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid X-API-KEY.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Messages — Actions" ],
        "summary" : "Move message to trash",
        "description" : "Moves the message to the trash (soft delete). For permanent deletion the trash needs to be emptied.",
        "operationId" : "deleteMessage",
        "parameters" : [ {
          "name" : "stableId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "maxLength" : 128,
            "minLength" : 0
          }
        } ],
        "responses" : {
          "503" : {
            "description" : "IMAP/SMTP server unavailable (MAIL_CONNECTION_ERROR).",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "204" : {
            "description" : "Message moved to trash."
          },
          "401" : {
            "description" : "Missing or invalid X-API-KEY.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/{stableId}/reply" : {
      "get" : {
        "tags" : [ "Messages — Write" ],
        "summary" : "Prepare reply (reply / reply-all)",
        "description" : "Returns a pre-filled MailRequest for replying to the given message. Parameter all=true = reply-all (reply to every recipient).",
        "operationId" : "prepareReply",
        "parameters" : [ {
          "name" : "stableId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "maxLength" : 128,
            "minLength" : 0
          }
        }, {
          "name" : "all",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "503" : {
            "description" : "IMAP/SMTP server unavailable (MAIL_CONNECTION_ERROR).",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MailRequest"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid X-API-KEY.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/messages/{stableId}/forward" : {
      "get" : {
        "tags" : [ "Messages — Write" ],
        "summary" : "Prepare forward",
        "description" : "Returns a pre-filled MailRequest for forwarding the message, including the quoted body and references to the original attachments.",
        "operationId" : "prepareForward",
        "parameters" : [ {
          "name" : "stableId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "maxLength" : 128,
            "minLength" : 0
          }
        } ],
        "responses" : {
          "503" : {
            "description" : "IMAP/SMTP server unavailable (MAIL_CONNECTION_ERROR).",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MailRequest"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid X-API-KEY.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ProblemDetail" : {
        "description" : "RFC 9457 problem+json — format of all error responses.",
        "properties" : {
          "type" : {
            "type" : "string",
            "format" : "uri",
            "description" : "URI identifier of the error type (e.g. .../errors/account_not_found)."
          },
          "title" : {
            "type" : "string",
            "description" : "Short human-readable description of the error class."
          },
          "status" : {
            "description" : "HTTP status code."
          },
          "detail" : {
            "type" : "string",
            "description" : "Error detail intended for the user."
          },
          "instance" : {
            "type" : "string",
            "format" : "uri",
            "description" : "URI of the specific occurrence (typically the request URI)."
          },
          "errorCode" : {
            "type" : "string",
            "description" : "Machine-readable error code (e.g. ACCOUNT_NOT_FOUND, VALIDATION_ERROR)."
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "Time when the error occurred (ISO-8601)."
          }
        }
      },
...
      "AttachmentRequest" : {
        "type" : "object",
        "properties" : {
          "fileName" : {
            "type" : "string",
            "maxLength" : 255,
            "minLength" : 0
          },
          "contentType" : {
            "type" : "string",
            "maxLength" : 255,
            "minLength" : 0
          },
          "base64Data" : {
            "type" : "string",
            "maxLength" : 73400320,
            "minLength" : 0
          }
        },
        "required" : [ "base64Data", "contentType", "fileName" ]
      },
      "MailRequest" : {
        "type" : "object",
        "properties" : {
          "to" : {
            "type" : "string",
            "minLength" : 1
          },
          "cc" : {
            "type" : "string"
          },
          "bcc" : {
            "type" : "string"
          },
          "subject" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "body" : {
            "type" : "string",
            "minLength" : 1
          },
          "attachments" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AttachmentRequest"
            }
          },
          "inReplyTo" : {
            "type" : "string"
          },
          "references" : {
            "type" : "string"
          }
        },
        "required" : [ "body", "subject", "to" ]
      },
      "SendAcceptedResponse" : {
        "type" : "object",
        "properties" : {
          "sendId" : {
            "type" : "string",
            "description" : "Correlation id for tracking the async send outcome over the notification stream.",
            "example" : "b3f1c2de-4a5b-6c7d-8e9f-0a1b2c3d4e5f"
          }
        }
      },
...
      "AttachmentResponse" : {
        "type" : "object",
        "properties" : {
          "partPath" : {
            "type" : "string"
          },
          "fileName" : {
            "type" : "string"
          },
          "contentType" : {
            "type" : "string"
          },
          "size" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "MailDetailResponse" : {
        "type" : "object",
        "properties" : {
          "stableId" : {
            "type" : "string",
            "description" : "Stable message identifier for the REST API. The client sends it to the detail, content and action endpoints."
          },
          "subject" : {
            "type" : "string"
          },
          "sender" : {
            "type" : "string"
          },
          "recipientsTo" : {
            "type" : "string"
          },
          "recipientsCc" : {
            "type" : "string"
          },
          "body" : {
            "type" : "string"
          },
          "receivedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "seen" : {
            "type" : "boolean"
          },
          "flagged" : {
            "type" : "boolean"
          },
          "answered" : {
            "type" : "boolean"
          },
          "messageId" : {
            "type" : "string"
          },
          "inReplyTo" : {
            "type" : "string"
          },
          "references" : {
            "type" : "string"
          },
          "hasAttachments" : {
            "type" : "boolean"
          },
          "attachments" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AttachmentResponse"
            }
          },
          "contentError" : {
            "type" : "string"
          },
          "threadId" : {
            "type" : "string",
            "description" : "Conversation identifier shared by every message of the same thread. Null only during the V2 backfill window."
          }
        }
      },
...
    },
    "securitySchemes" : {
      "ApiKeyAuth" : {
        "type" : "apiKey",
        "description" : "Internal API key that the desktop client loads from session.json and sends in the X-API-KEY header.",
        "name" : "X-API-KEY",
        "in" : "header"
      }
    }
  }
}